requirements.txt
Pillow==7.2.0
pyTelegramBotAPI==3.7.3
python-dotenv==0.14.0
